+++ sdk/clientcore/core/src/main/java/io/clientcore/core/http/pipeline/KeyCredentialPolicy.java
@@ -74,8 +74,16 @@ public class KeyCredentialPolicy extends HttpCredentialPolicy {
         this.prefix = prefix != null ? prefix.trim() : null;
     }
 
+    /**
+     * {@inheritDoc}
+     * @throws IllegalStateException If the request is not using {@code HTTPS}.
+     */
     @Override
     public Response<BinaryData> process(HttpRequest httpRequest, HttpPipelineNextPolicy next) {
+        if (!"https".equals(httpRequest.getUri().getScheme())) {
+            throw LOGGER.logThrowableAsError(
+                new IllegalStateException("Key credentials require HTTPS to prevent leaking the key."));
+        }
         setCredential(httpRequest.getHeaders());
         return next.process();
     }